MODORDERS-1105. Refactor to introduce new service for updating statuses
