+++ src/main/java/org/folio/helper/PurchaseOrderLineHelper.java
@@ -79,7 +79,7 @@ import org.folio.service.finance.transaction.EncumbranceService;
 import org.folio.service.finance.transaction.EncumbranceWorkflowStrategy;
 import org.folio.service.finance.transaction.EncumbranceWorkflowStrategyFactory;
 import org.folio.service.inventory.InventoryInstanceManager;
-import org.folio.service.inventory.InventoryItemManager;
+import org.folio.service.inventory.InventoryItemStatusSyncService;
 import org.folio.service.invoice.POLInvoiceLineRelationService;
 import org.folio.service.orders.CompositePoLineValidationService;
 import org.folio.service.orders.OrderInvoiceRelationService;
@@ -108,7 +108,7 @@ public class PurchaseOrderLineHelper {
   private static final String OTHER = "other";
   private static final String QUERY_BY_PO_LINE_ID = "poLineId==";
 
-  private final InventoryItemManager inventoryItemManager;
+  private final InventoryItemStatusSyncService itemStatusSyncService;
   private final InventoryInstanceManager inventoryInstanceManager;
   private final EncumbranceService encumbranceService;
   private final ExpenseClassValidationService expenseClassValidationService;
@@ -123,7 +123,7 @@ public class PurchaseOrderLineHelper {
   private final OrganizationService organizationService;
   private final POLInvoiceLineRelationService polInvoiceLineRelationService;
 
-  public PurchaseOrderLineHelper(InventoryItemManager inventoryItemManager,
+  public PurchaseOrderLineHelper(InventoryItemStatusSyncService inventoryItemStatusSyncService,
                                  InventoryInstanceManager inventoryInstanceManager,
                                  EncumbranceService encumbranceService,
                                  ExpenseClassValidationService expenseClassValidationService,
@@ -138,7 +138,7 @@ public class PurchaseOrderLineHelper {
                                  POLInvoiceLineRelationService polInvoiceLineRelationService,
                                  OrganizationService organizationService) {
 
-    this.inventoryItemManager = inventoryItemManager;
+    this.itemStatusSyncService = inventoryItemStatusSyncService;
     this.inventoryInstanceManager = inventoryInstanceManager;
     this.encumbranceService = encumbranceService;
     this.expenseClassValidationService = expenseClassValidationService;
@@ -745,29 +745,11 @@ public class PurchaseOrderLineHelper {
             logger.info("updateInventoryItemStatus:: not canceled po lines not found, returning...");
             return Future.succeededFuture(null);
           }
-          return GenericCompositeFuture.all(
-            PoLineCommonUtil.getTenantsFromLocations(compOrderLine)
-              .stream()
-              .map(tenantId -> updateItemStatusForTenant(compOrderLine, RequestContextUtil.createContextWithNewTenantId(requestContext, tenantId)))
-              .toList()
-          ).mapEmpty();
+          return itemStatusSyncService.updateInventoryItemStatus(compOrderLine.getId(), compOrderLine.getLocations(),
+            ItemStatus.ORDER_CLOSED, requestContext);
         });
     }
     return Future.succeededFuture(null);
-  }
-
-  private Future<Void> updateItemStatusForTenant(CompositePoLine compOrderLine, RequestContext requestContext) {
-    return inventoryItemManager.getItemsByPoLineIdsAndStatus(List.of(compOrderLine.getId()), ItemStatus.ON_ORDER.value(), requestContext)
-      .compose(items -> {
-        //Each poLine can have only one linked item
-        Optional<JsonObject> poLineItem = items.stream()
-          .filter(item -> compOrderLine.getId().equals(item.getString("purchaseOrderLineIdentifier"))).findFirst();
-        if (poLineItem.isPresent()) {
-          JsonObject updatedItem = updateItemStatus(poLineItem.get(), ItemStatus.ORDER_CLOSED);
-          return inventoryItemManager.updateItem(updatedItem, requestContext);
-        }
-        return Future.succeededFuture(null);
-      });
   }
 
   private boolean isStatusChanged(CompositePoLine compOrderLine, PoLine lineFromStorage) {
@@ -783,14 +765,6 @@ public class PurchaseOrderLineHelper {
   private boolean isCurrentStatusCanceled(CompositePoLine compOrderLine) {
     return CompositePoLine.ReceiptStatus.CANCELLED.equals(compOrderLine.getReceiptStatus()) ||
       CompositePoLine.PaymentStatus.CANCELLED.equals(compOrderLine.getPaymentStatus());
-  }
-
-  private JsonObject updateItemStatus(JsonObject poLineItem, ItemStatus itemStatus) {
-    JsonObject status = new JsonObject();
-    status.put("name", itemStatus);
-    status.put("date", Instant.now());
-    poLineItem.put("status", status);
-    return poLineItem;
   }
 
   private boolean isEncumbranceUpdateNeeded(CompositePurchaseOrder compOrder, CompositePoLine compositePoLine, PoLine storagePoLine) {
