+++ src/main/java/org/folio/orders/utils/RequestContextUtil.java
@@ -1,10 +1,14 @@
 package org.folio.orders.utils;
 
 import org.apache.commons.collections4.map.CaseInsensitiveMap;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+import org.folio.helper.PurchaseOrderHelper;
 import org.folio.okapi.common.XOkapiHeaders;
 import org.folio.rest.core.models.RequestContext;
 
 public class RequestContextUtil {
+  private static final Logger logger = LogManager.getLogger(PurchaseOrderHelper.class);
 
   private RequestContextUtil() {
   }
@@ -15,6 +19,7 @@ public class RequestContextUtil {
     }
     var modifiedHeaders = new CaseInsensitiveMap<>(requestContext.getHeaders());
     modifiedHeaders.put(XOkapiHeaders.TENANT, tenantId);
+    logger.info("Request context has been changes with new tenant: {}", tenantId);
     return new RequestContext(requestContext.getContext(), modifiedHeaders);
   }
